새로고침 하지 않으면 작동하지 않는 버그 해결
1. 로그아웃 시 직접 새로 고침을 해야 로그인이 가능한 버그 수정

2. 매치 삭제했을 시 직접 새로 고침을 해야 반영이 되던 버그 수정
