+++ src/components/MyTeamMatchList.jsx
@@ -1,21 +1,11 @@
 // MainList.jsx
-import React from 'react';
+import React, { useState } from 'react';
 import { Card, Group, Text, Badge, Box, Button } from '@mantine/core';
 import { useNavigate } from 'react-router-dom';
-import { deleteMatch } from '../api/main';
 
-const MyTeamMatchList = ({ matches }) => {
+
+const MyTeamMatchList = ({ matches, onDelete }) => {
   const navigate = useNavigate();
-
-  const handleDelete = async (matchId) => {
-    try {
-      await deleteMatch(matchId);
-      alert('매치가 삭제되었습니다.');
-      // 페이지를 새로고침하거나, 상태를 업데이트하여 삭제된 매치를 제거하세요.
-    } catch (error) {
-      alert('매치를 삭제할 수 있는 권한이 없습니다.');
-    }
-  };
 
   return (
     <>
@@ -49,7 +39,7 @@ const MyTeamMatchList = ({ matches }) => {
               gradient={{ from: 'green', to: 'lime', deg: 64 }}>
               신청 정보 보기
             </Button>
-            <Button onClick={() => handleDelete(match.id)} variant="gradient"
+            <Button onClick={() => onDelete(match.id)} variant="gradient"
               gradient={{ from: 'red', to: 'pink', deg: 131 }} style={{ marginLeft: '10px' }}>
               삭제하기
             </Button>
